feat(user): account details modal
Add account details modal to edit the profile.
The button to open the modal reveals when the user is visiting its profile.
Removed the old profile modal, the profile image now navigates to the profile page.
