+++ __tests__/AccountDetailsModal.tsx
@@ -1,17 +1,17 @@
 import React from 'react';
-import { render, RenderResult, screen, waitFor } from '@testing-library/react';
-import { mocked } from 'ts-jest/utils';
-import ProfileModal, { Props } from '../components/ProfileModal';
-import AuthContext from '../components/AuthContext';
 import { LoggedUser, updateProfile } from '../lib/user';
+import AccountDetailsModal from '../components/AccountDetailsModal';
+import { render, RenderResult, screen, waitFor } from '@testing-library/react';
+import AuthContext from '../components/AuthContext';
+import { mocked } from 'ts-jest/utils';
 
 jest.mock('../lib/user', () => ({
   ...jest.requireActual('../lib/user'),
   updateProfile: jest.fn(),
 }));
 
-const profiledUpdated = jest.fn();
 const logout = jest.fn();
+const onRequestClose = jest.fn();
 
 beforeEach(() => {
   jest.resetAllMocks();
@@ -25,74 +25,65 @@ const defaultUser = {
   image: 'https://daily.dev/ido.png',
   infoConfirmed: true,
   premium: false,
-  createdAt: '',
+  createdAt: '2020-07-26T13:04:35.000Z',
 };
 
-const renderComponent = (
-  props: Partial<Props> = {},
-  user: Partial<LoggedUser> = {},
-): RenderResult => {
-  const defaultProps: Props = {
-    isOpen: true,
-    profiledUpdated,
-    ariaHideApp: false,
-    confirmMode: false,
-  };
-
+const renderComponent = (user: Partial<LoggedUser> = {}): RenderResult => {
   return render(
     <AuthContext.Provider
       value={{
         user: { ...defaultUser, ...user },
         shouldShowLogin: false,
         showLogin: jest.fn(),
-        showProfile: jest.fn(),
+        updateUser: jest.fn(),
         logout,
       }}
     >
-      <ProfileModal {...defaultProps} {...props} />
+      <AccountDetailsModal
+        isOpen={true}
+        onRequestClose={onRequestClose}
+        ariaHideApp={false}
+      />
     </AuthContext.Provider>,
   );
 };
 
-it('should enable submit when form is valid', () => {
+it('should show profile image', () => {
+  renderComponent();
+  const el = screen.getByAltText('Your profile image');
+  expect(el).toHaveAttribute('src', defaultUser.image);
+});
+
+it('should disable submit when form is invalid', () => {
   renderComponent();
   const el = screen.getByText('Save changes');
-  expect(el.getAttribute('disabled')).toBeFalsy();
+  expect(el).toBeDisabled();
+});
+
+it('should enable submit when form is valid', () => {
+  renderComponent({ username: 'idoshamun' });
+  const el = screen.getByText('Save changes');
+  expect(el).toBeEnabled();
 });
 
 it('should submit information on button click', async () => {
-  renderComponent();
+  renderComponent({ username: 'idoshamun' });
   mocked(updateProfile).mockResolvedValue(defaultUser);
   screen.getByText('Save changes').click();
-  await waitFor(() => expect(profiledUpdated).toBeCalledWith(defaultUser));
+  await waitFor(() => expect(updateProfile).toBeCalledTimes(1));
   expect(updateProfile).toBeCalledWith({
     name: 'Ido Shamun',
     email: 'ido@acme.com',
+    username: 'idoshamun',
     company: null,
     title: null,
     acceptedMarketing: false,
+    bio: null,
+    github: null,
+    portfolio: null,
+    twitter: null,
   });
-});
-
-it('should show server error', async () => {
-  renderComponent();
-  mocked(updateProfile).mockResolvedValue({
-    error: true,
-    code: 1,
-    message: '',
-  });
-  screen.getByText('Save changes').click();
-  await waitFor(() =>
-    expect(updateProfile).toBeCalledWith({
-      name: 'Ido Shamun',
-      email: 'ido@acme.com',
-      company: null,
-      title: null,
-      acceptedMarketing: false,
-    }),
-  );
-  expect(profiledUpdated).toBeCalledTimes(0);
-  expect(screen.getByRole('alert')).toBeInTheDocument();
+  expect(onRequestClose).toBeCalledTimes(1);
 });
 
 it('should logout on button click', async () => {
