+++ __tests__/CommentActionButtons.tsx
@@ -67,8 +67,8 @@ const renderComponent = (
           user,
           shouldShowLogin: false,
           showLogin,
-          showProfile: jest.fn(),
           logout: jest.fn(),
+          updateUser: jest.fn(),
         }}
       >
         <CommentActionButtons {...props} />
