+++ __tests__/MainComment.tsx
@@ -56,8 +56,8 @@ const renderLayout = (
           user,
           shouldShowLogin: false,
           showLogin: jest.fn(),
-          showProfile: jest.fn(),
           logout: jest.fn(),
+          updateUser: jest.fn(),
         }}
       >
         <MainComment {...defaultProps} {...props} />
