+++ __tests__/SubComment.tsx
@@ -56,8 +56,8 @@ const renderLayout = (
           user,
           shouldShowLogin: false,
           showLogin: jest.fn(),
-          showProfile: jest.fn(),
           logout: jest.fn(),
+          updateUser: jest.fn(),
         }}
       >
         <SubComment {...defaultProps} {...props} />
