+++ lib/constants.ts
@@ -1,3 +1,8 @@
+export const faq = 'https://github.com/dailydotdev/daily/blob/master/FAQs.md';
+export const requestFeature =
+  'https://github.com/dailydotdev/daily/issues/new?assignees=&labels=Type%3A+Feature&template=---feature-request.md&title=%F0%9F%A7%A9+FEATURE+REQUEST%3A+';
+export const reportIssue =
+  'https://github.com/dailydotdev/daily/issues/new?assignees=&labels=Type%3A+Bug&template=---bug-report.md&title=%F0%9F%90%9B+BUG%3A+';
 export const termsOfService = 'https://daily.dev/tos';
 export const privacyPolicy = 'https://daily.dev/privacy';
 export const cookiePolicy = 'https://daily.dev/cookies';
