+++ package.json
@@ -30,6 +30,7 @@
     "react-modal": "^3.11.2",
     "styled-components": "^5.1.1",
     "styled-normalize": "^8.0.7",
+    "swr": "^0.3.4",
     "universal-cookie": "^4.0.3"
   },
   "devDependencies": {
